+++ R/create.clone.genome.distribution.densityplot.R
@@ -8,7 +8,7 @@ create.clone.genome.distribution.densityplot <- function(
 
     return(BoutrosLab.plotting.general::create.scatterplot(
         filename = save.plt,
-        formula = count ~ x,
+        formula = y ~ x,
         data = density.df,
         groups = density.df$clone.id,
         xlab.label = 'Chromosome',
@@ -23,18 +23,4 @@ create.clone.genome.distribution.densityplot <- function(
         abline.lwd = 1.2,
         ...
         ));
-    }
-
-calculate.density.and.scale <- function(cluster.df) {
-    # density should be generated using unque SNV count
-    density <- density(
-        x = cluster.df$genome.pos,
-        bw = 'nrd',
-        adjust = 0.05, # set to 1E9/3E9 to get density per megabase
-        na.rm = TRUE);
-    density.df <- as.data.frame(density[c('x','y')]);
-    density.df$clone.id <- unique(cluster.df$clone.id);
-    density.df$count <- nrow(cluster.df) / sum(density.df$y) * density.df$y;
-
-    return(density.df)
     }