+++ NEWS.md
@@ -13,13 +13,17 @@
 * Wrapper function for `SRCgrob` to automatically save plots to file
 * Add option to annotate the CCF summary heatmap with the cell values.
 * `plotting.direction` argument to `SRCGrob` to support arbitrary direction of tree "flow". Can be one of 'down' (default), 'right', 'up', 'left' or any number (in degrees).
+* Function to generate single-sample density plot
+* Add support for 1xn and 1x1 heatmaps.
+* Add `get.colours.in.order` function to get a list of colours and corresponding clone ID order.
+* Add `sample.order` and `clone.order` as input parameters to `create.cluster.heatmap`
 
 ## Update
 * Fixed angle calculation bug where child angles do not follow
   their parent angle, instead moving "downward" at 0 degrees.
 * Updated package metadata and README
 * Split angle handling for `radial` and `dendrogram` modes to optimized each
-* Set default parameters for heatmaps, defaulting too BPG defaults unless necessary
+* Fixed bug where horizontal tree segments rendered as zero-length segments
 * Updated changelog format to NEWS.md Markdown format
 * Refactored use of plyr/dplyr and stringr functions to remove dependencies
 * Set default parameters for heatmaps, defaulting to BPG defaults unless necessary
@@ -27,10 +31,13 @@
 * Define color scheme with a single value, automatically generating other shades.
 * Report the number of SNVs per clone in the legend for the clone-genome distribution plot.
 * Updated the user guide to reflect new features.
+* Fix bug where the x-axis only renders when y-axis is also rendered.
+* Fix issue when creating polygons with more than 2 siblings
 
 ## Bug
 * Resolved issue where the spread parameter was not applied in dendrogram mode.
 * Resolved issue for simple dendrogram trees ( < 6 nodes or binary tree), where node angles were not calculated correctly.
+
 
 # CancerEvolutionVisualization 2.0.1 (2023-11-17)